fix diff mode truncation to first 512 byte
when checking the return code of fwrite, make sure we compare bytes with bytes,
not one "element" rc bytes in size.

Embarrassingly enough, I introduced the bug in both rsync.c and update.c
with 6f5ba5e, but then fixed only the update.c with f177b72,
and overlooked this one.

Thanks to moko-als on github for point that out.