try to avoid (temporary) -rw------- root:root files on receiving side
The "temporary" newfname typically is created as "-rw------- root:root".
Before we rename it into place, "clone" the permissions and ownership of
the old file to avoid files already with the "correct name", but root
owned and unreadable by the applications, if even for a short time.

If the old file does not exist (or cannot be stat()ed), ignore.

If we cannot rename into place, but have to truncate/rewrite,
we keep the existing file anyways, so no ownership/permission change
happens.

If permissions and/or ownership has also changed, in addition to
content, those will be explicitly synced "soon".

If there are errors while cloning the ownership or permissions, so what?
Ignore, it stays at root:root 0600, and won't be worse than before.

Can we do better?
Until we change the ugly ad-hoc csync2 "protocol" to use "transaction"
like semantics, and know all of the changed content, plus ownership,
permissions, mtime, ACLs and other meta data as context information
before starting to act on it on the receiving side, I don't see how.

Any volunteers for implementing "transactions", or at least a mode
where we can push ownership, permissions, mtime and content (-diff)
in *one* "protocol command"?